requirements.txt
Flask==2.0.2
numpy==1.21.4
imageio
azure-storage-blob
onnxruntime